Configure updateme lookup of new package versions

This function is a helper for setting the "updateme.sources" global option. It provides a user-friendly interface and validation of the options you set.

Usage

updateme_sources_set(...)

Arguments

...

Named or unnamed arguments. Values should be either:

"github"/"gitlab": new versions will be found from the package development version on GitHub/GitLab, if a repo can be identified using the package DESCRIPTION

One of names(getOption("repo")): latest versions will be taken from this source, if available

A URL pointing to a GitHub/GitLab repo, e.g. "https://github.com/wurli/updateme": the latest version for this particular package will be taken from this project

NULL: updateme will not attempt to query new versions. Note that NULL inputs must always be named (i.e. you must specify this 'per package').

If arguments are named, names should indicate package which the option should apply to. If unnamed, the option will apply to all packages. See examples for more information.

Value

The result of setting options(updateme.sources = <new_options>)

Private Repositories

updateme supports packages stored in private repositories on GitHub and GitLab. To get upstream package version from either, you should only have to configure a personal access token (PAT).

For GitHub packages, updateme checks, in order:

The UPDATEME_GITHUB_PAT environmental variable

Any personal access tokens configured using gitcreds::gitcreds_set()

The GITHUB_PAT environmental variable

The GITHUB_TOKEN environmental variable

For GitLab packages, updateme checks, in order:

The UPDATEME_GITLAB_PAT environmental variable

Any personal access tokens configured using gitcreds::gitcreds_set()

The GITLAB_PAT environmental variable

The GITLAB_TOKEN environmental variable

See also

Examples

if (FALSE) {

  # If you want to check non-standard repos for new versions of packages,
  # you'll first have to set the repos global option. Note that each
  # option must be named for compatibility with {updateme}
  options(repos = c(

    # Your default repos, e.g. c(CRAN = "https://cloud.r-project.org")
    getOption("repos"),

    # The tidyverse r-universe, including dev versions of tidyverse packages
    tidyverse = "https://tidyverse.r-universe.dev",

    # The r-lib r-universe, including dev versions of infrastructure packages
    # like {cli}, {rlang}, etc
    rlib = "https://r-lib.r-universe.dev"
  ))

  # 1. New versions will first be looked up in the tidyverse r-universe
  # 2. If not found, they will be looked up from your usual CRAN mirror
  # 3. {bslib} will always be first looked up from GitHub
  # 4. {cli} will always be first looked up from the r-lib r-universe
  updateme_sources_set(
    "tidyverse",
    "CRAN",
    bslib = "https://github.com/rstudio/bslib", # Name is optional here
    cli = "rlib"
  )
}